Move TypeSystemLanguageSupportTests
